for gui: cleaned up gui.py, added gui components file
